+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [1.5.4] - 2017-07-31
 ### Added
+- Added support for allowing job to specify which executor (cook|mesos) to use, from @shamsimam
 - Added support for throttling rate of publishing instance progress updates, from @shamsimam
 - Added authorization check for job creation, from @dposada